+++ src/kmers.hpp
@@ -63,7 +63,7 @@ extern "C" {
 	void mark_kmer_used(const uint64_t kmer_int, const hash_map *hm, const int tpl_id, const int locus);
 	int kmer_is_used(const uint64_t kmer_int, hash_map *hm);
 	void read_tpl_using_kmer(const uint64_t kmer_int, const hash_map *hm,
-			int *tpl_id, int *locus, uint64_t *value);
+			uint64_t *tpl_id, int *locus, uint64_t *value);
 	void kmer_aln_query(const bwa_seq_t *query, const hash_map *hm, GPtrArray *hits);
 
 #ifdef __cplusplus
